refactor: share store initial state
